Tests for two_state_tps.py
